+++ sdk/textanalytics/azure-ai-textanalytics/src/main/java/com/azure/ai/textanalytics/implementation/HealthcareTaskResultPropertiesHelper.java
@@ -1,0 +1,39 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+package com.azure.ai.textanalytics.implementation;
+
+import com.azure.ai.textanalytics.models.HealthcareTaskResult;
+import com.azure.ai.textanalytics.util.RecognizeHealthcareEntitiesResultCollection;
+
+/**
+ * The helper class to set the non-public properties of an {@link HealthcareTaskResult} instance.
+ */
+public final class HealthcareTaskResultPropertiesHelper {
+    private static HealthcareTaskResultAccessor accessor;
+
+    private HealthcareTaskResultPropertiesHelper() { }
+
+    /**
+     * Type defining the methods to set the non-public properties of an {@link HealthcareTaskResult} instance.
+     */
+    public interface HealthcareTaskResultAccessor {
+        void setRecognizeHealthcareEntitiesResultCollection(HealthcareTaskResult healthcareTaskResult,
+            RecognizeHealthcareEntitiesResultCollection recognizeHealthcareEntitiesResultCollection);
+    }
+
+    /**
+     * The method called from {@link HealthcareTaskResult} to set it's accessor.
+     *
+     * @param healthcareTaskResultAccessor The accessor.
+     */
+    public static void setAccessor(final HealthcareTaskResultAccessor healthcareTaskResultAccessor) {
+        accessor = healthcareTaskResultAccessor;
+    }
+
+    public static void setResult(HealthcareTaskResult healthcareTaskResult,
+        RecognizeHealthcareEntitiesResultCollection recognizeHealthcareEntitiesResultCollection) {
+        accessor.setRecognizeHealthcareEntitiesResultCollection(healthcareTaskResult,
+            recognizeHealthcareEntitiesResultCollection);
+    }
+}