+++ sdk/textanalytics/azure-ai-textanalytics/src/test/java/com/azure/ai/textanalytics/TestUtils.java
@@ -3,6 +3,12 @@
 
 package com.azure.ai.textanalytics;
 
+import com.azure.ai.textanalytics.implementation.HealthcareEntityCollectionPropertiesHelper;
+import com.azure.ai.textanalytics.implementation.HealthcareEntityPropertiesHelper;
+import com.azure.ai.textanalytics.implementation.HealthcareEntityRelationPropertiesHelper;
+import com.azure.ai.textanalytics.implementation.HealthcareTaskResultPropertiesHelper;
+import com.azure.ai.textanalytics.implementation.RecognizeHealthcareEntitiesResultCollectionPropertiesHelper;
+import com.azure.ai.textanalytics.implementation.RecognizeHealthcareEntitiesResultPropertiesHelper;
 import com.azure.ai.textanalytics.models.AnalyzeSentimentResult;
 import com.azure.ai.textanalytics.models.AspectSentiment;
 import com.azure.ai.textanalytics.models.CategorizedEntity;
@@ -13,6 +19,10 @@ import com.azure.ai.textanalytics.models.DetectedLanguage;
 import com.azure.ai.textanalytics.models.DocumentSentiment;
 import com.azure.ai.textanalytics.models.EntityCategory;
 import com.azure.ai.textanalytics.models.ExtractKeyPhraseResult;
+import com.azure.ai.textanalytics.models.HealthcareEntity;
+import com.azure.ai.textanalytics.models.HealthcareEntityCollection;
+import com.azure.ai.textanalytics.models.HealthcareEntityRelation;
+import com.azure.ai.textanalytics.models.HealthcareTaskResult;
 import com.azure.ai.textanalytics.models.KeyPhrasesCollection;
 import com.azure.ai.textanalytics.models.LinkedEntity;
 import com.azure.ai.textanalytics.models.LinkedEntityCollection;
@@ -22,6 +32,7 @@ import com.azure.ai.textanalytics.models.OpinionSentiment;
 import com.azure.ai.textanalytics.models.PiiEntity;
 import com.azure.ai.textanalytics.models.PiiEntityCollection;
 import com.azure.ai.textanalytics.models.RecognizeEntitiesResult;
+import com.azure.ai.textanalytics.models.RecognizeHealthcareEntitiesResult;
 import com.azure.ai.textanalytics.models.RecognizeLinkedEntitiesResult;
 import com.azure.ai.textanalytics.models.RecognizePiiEntitiesResult;
 import com.azure.ai.textanalytics.models.SentenceSentiment;
@@ -34,6 +45,7 @@ import com.azure.ai.textanalytics.util.AnalyzeSentimentResultCollection;
 import com.azure.ai.textanalytics.util.DetectLanguageResultCollection;
 import com.azure.ai.textanalytics.util.ExtractKeyPhrasesResultCollection;
 import com.azure.ai.textanalytics.util.RecognizeEntitiesResultCollection;
+import com.azure.ai.textanalytics.util.RecognizeHealthcareEntitiesResultCollection;
 import com.azure.ai.textanalytics.util.RecognizeLinkedEntitiesResultCollection;
 import com.azure.ai.textanalytics.util.RecognizePiiEntitiesResultCollection;
 import com.azure.core.exception.HttpResponseException;
@@ -95,6 +107,11 @@ final class TestUtils {
 
     static final String PII_ENTITY_OFFSET_INPUT = "SSN: 859-98-0987";
     static final String SENTIMENT_OFFSET_INPUT = "The hotel was dark and unclean.";
+
+    static final List<String> HEALTHCARE_INPUTS = asList(
+        "The patient is a 54-year-old gentleman with a history of progressive angina over the past several months.",
+        "The patient went for six minutes with minimal ST depressions in the anterior lateral leads , thought due to fatigue and wrist pain , his anginal equivalent.");
+
 
     // "personal" and "social" are common to both English and Spanish and if given with limited context the
     // response will be based on the "US" country hint. If the origin of the text is known to be coming from
@@ -415,6 +432,189 @@ final class TestUtils {
                     36
                 )
             )), null);
+    }
+
+    /**
+     * Helper method that get the expected HealthcareTaskResult result.
+     */
+    static HealthcareTaskResult getExpectedHealthcareTaskResult() {
+        final HealthcareTaskResult healthcareTaskResult = new HealthcareTaskResult(null, null,
+            null, null, null, null);
+        HealthcareTaskResultPropertiesHelper.setResult(healthcareTaskResult, getExpectedBatchHealthcareEntities());
+        return healthcareTaskResult;
+    }
+
+    /**
+     * Helper method that get the expected RecognizeHealthcareEntitiesResultCollection result.
+     */
+    static RecognizeHealthcareEntitiesResultCollection getExpectedBatchHealthcareEntities() {
+        TextDocumentBatchStatistics textDocumentBatchStatistics = new TextDocumentBatchStatistics(2, 2, 0, 2);
+        TextDocumentStatistics textDocumentStatistics1 = new TextDocumentStatistics(105, 1);
+        TextDocumentStatistics textDocumentStatistics2 = new TextDocumentStatistics(156, 1);
+
+        final HealthcareEntity healthcareEntity1 = new HealthcareEntity();
+        HealthcareEntityPropertiesHelper.setText(healthcareEntity1, "54-year-old");
+        HealthcareEntityPropertiesHelper.setCategory(healthcareEntity1, EntityCategory.fromString("Age"));
+        HealthcareEntityPropertiesHelper.setSubcategory(healthcareEntity1, null);
+        HealthcareEntityPropertiesHelper.setConfidenceScore(healthcareEntity1, 1.0);
+        HealthcareEntityPropertiesHelper.setOffset(healthcareEntity1, 17);
+        HealthcareEntityPropertiesHelper.setNegated(healthcareEntity1, false);
+
+        final HealthcareEntity healthcareEntity2 = new HealthcareEntity();
+        HealthcareEntityPropertiesHelper.setText(healthcareEntity2, "gentleman");
+        HealthcareEntityPropertiesHelper.setCategory(healthcareEntity2, EntityCategory.fromString("Gender"));
+        HealthcareEntityPropertiesHelper.setSubcategory(healthcareEntity2, null);
+        HealthcareEntityPropertiesHelper.setConfidenceScore(healthcareEntity2, 1.0);
+        HealthcareEntityPropertiesHelper.setOffset(healthcareEntity2, 29);
+        HealthcareEntityPropertiesHelper.setNegated(healthcareEntity2, false);
+        // there are too many entity links, we can just assert it is not null.
+        HealthcareEntityPropertiesHelper.setHealthcareEntityLinks(healthcareEntity2, new ArrayList<>());
+
+        final HealthcareEntity healthcareEntity3 = new HealthcareEntity();
+        HealthcareEntityPropertiesHelper.setText(healthcareEntity3, "progressive");
+        HealthcareEntityPropertiesHelper.setCategory(healthcareEntity3, EntityCategory.fromString("ConditionQualifier"));
+        HealthcareEntityPropertiesHelper.setSubcategory(healthcareEntity3, null);
+        HealthcareEntityPropertiesHelper.setConfidenceScore(healthcareEntity3, 0.91);
+        HealthcareEntityPropertiesHelper.setOffset(healthcareEntity3, 57);
+        HealthcareEntityPropertiesHelper.setNegated(healthcareEntity3, false);
+
+        final HealthcareEntity healthcareEntity4 = new HealthcareEntity();
+        HealthcareEntityPropertiesHelper.setText(healthcareEntity4, "angina");
+        HealthcareEntityPropertiesHelper.setCategory(healthcareEntity4, EntityCategory.fromString("SymptomOrSign"));
+        HealthcareEntityPropertiesHelper.setSubcategory(healthcareEntity4, null);
+        HealthcareEntityPropertiesHelper.setConfidenceScore(healthcareEntity4, 0.81);
+        HealthcareEntityPropertiesHelper.setOffset(healthcareEntity4, 69);
+        HealthcareEntityPropertiesHelper.setNegated(healthcareEntity4, false);
+        // there are too many entity links, we can just assert it is not null.
+        HealthcareEntityPropertiesHelper.setHealthcareEntityLinks(healthcareEntity4, new ArrayList<>());
+
+        final HealthcareEntity healthcareEntity5 = new HealthcareEntity();
+        HealthcareEntityPropertiesHelper.setText(healthcareEntity5, "past several months");
+        HealthcareEntityPropertiesHelper.setCategory(healthcareEntity5, EntityCategory.fromString("Time"));
+        HealthcareEntityPropertiesHelper.setSubcategory(healthcareEntity5, null);
+        HealthcareEntityPropertiesHelper.setConfidenceScore(healthcareEntity5, 1.0);
+        HealthcareEntityPropertiesHelper.setOffset(healthcareEntity5, 85);
+        HealthcareEntityPropertiesHelper.setNegated(healthcareEntity5, false);
+
+        final HealthcareEntity healthcareEntity6 = new HealthcareEntity();
+        HealthcareEntityPropertiesHelper.setText(healthcareEntity6, "minutes");
+        HealthcareEntityPropertiesHelper.setCategory(healthcareEntity6, EntityCategory.fromString("Time"));
+        HealthcareEntityPropertiesHelper.setSubcategory(healthcareEntity6, null);
+        HealthcareEntityPropertiesHelper.setConfidenceScore(healthcareEntity6, 0.87);
+        HealthcareEntityPropertiesHelper.setOffset(healthcareEntity6, 25);
+        HealthcareEntityPropertiesHelper.setNegated(healthcareEntity6, false);
+
+        final HealthcareEntity healthcareEntity7 = new HealthcareEntity();
+        HealthcareEntityPropertiesHelper.setText(healthcareEntity7, "minimal");
+        HealthcareEntityPropertiesHelper.setCategory(healthcareEntity7, EntityCategory.fromString("ConditionQualifier"));
+        HealthcareEntityPropertiesHelper.setSubcategory(healthcareEntity7, null);
+        HealthcareEntityPropertiesHelper.setConfidenceScore(healthcareEntity7, 1.0);
+        HealthcareEntityPropertiesHelper.setOffset(healthcareEntity7, 38);
+        HealthcareEntityPropertiesHelper.setNegated(healthcareEntity7, false);
+
+        final HealthcareEntity healthcareEntity8 = new HealthcareEntity();
+        HealthcareEntityPropertiesHelper.setText(healthcareEntity8, "ST depressions in the anterior lateral leads");
+        HealthcareEntityPropertiesHelper.setCategory(healthcareEntity8, EntityCategory.fromString("SymptomOrSign"));
+        HealthcareEntityPropertiesHelper.setSubcategory(healthcareEntity8, null);
+        HealthcareEntityPropertiesHelper.setConfidenceScore(healthcareEntity8, 1.0);
+        HealthcareEntityPropertiesHelper.setOffset(healthcareEntity8, 46);
+        HealthcareEntityPropertiesHelper.setNegated(healthcareEntity8, false);
+
+        final HealthcareEntity healthcareEntity9 = new HealthcareEntity();
+        HealthcareEntityPropertiesHelper.setText(healthcareEntity9, "fatigue");
+        HealthcareEntityPropertiesHelper.setCategory(healthcareEntity9, EntityCategory.fromString("SymptomOrSign"));
+        HealthcareEntityPropertiesHelper.setSubcategory(healthcareEntity9, null);
+        HealthcareEntityPropertiesHelper.setConfidenceScore(healthcareEntity9, 1.0);
+        HealthcareEntityPropertiesHelper.setOffset(healthcareEntity9, 108);
+        HealthcareEntityPropertiesHelper.setNegated(healthcareEntity9, false);
+        // there are too many entity links, we can just assert it is not null.
+        HealthcareEntityPropertiesHelper.setHealthcareEntityLinks(healthcareEntity9, new ArrayList<>());
+
+        final HealthcareEntity healthcareEntity10 = new HealthcareEntity();
+        HealthcareEntityPropertiesHelper.setText(healthcareEntity10, "wrist pain");
+        HealthcareEntityPropertiesHelper.setCategory(healthcareEntity10, EntityCategory.fromString("SymptomOrSign"));
+        HealthcareEntityPropertiesHelper.setSubcategory(healthcareEntity10, null);
+        HealthcareEntityPropertiesHelper.setConfidenceScore(healthcareEntity10, 1.0);
+        HealthcareEntityPropertiesHelper.setOffset(healthcareEntity10, 120);
+        HealthcareEntityPropertiesHelper.setNegated(healthcareEntity10, false);
+        // there are too many entity links, we can just assert it is not null.
+        HealthcareEntityPropertiesHelper.setHealthcareEntityLinks(healthcareEntity10, new ArrayList<>());
+
+        final HealthcareEntity healthcareEntity11 = new HealthcareEntity();
+        HealthcareEntityPropertiesHelper.setText(healthcareEntity11, "anginal equivalent");
+        HealthcareEntityPropertiesHelper.setCategory(healthcareEntity11, EntityCategory.fromString("SymptomOrSign"));
+        HealthcareEntityPropertiesHelper.setSubcategory(healthcareEntity11, null);
+        HealthcareEntityPropertiesHelper.setConfidenceScore(healthcareEntity11, 1.0);
+        HealthcareEntityPropertiesHelper.setOffset(healthcareEntity11, 137);
+        HealthcareEntityPropertiesHelper.setNegated(healthcareEntity11, false);
+        // there are too many entity links, we can just assert it is not null.
+        HealthcareEntityPropertiesHelper.setHealthcareEntityLinks(healthcareEntity11, new ArrayList<>());
+
+        final HealthcareEntityRelation healthcareEntityRelation1 = new HealthcareEntityRelation();
+        HealthcareEntityRelationPropertiesHelper.setRelationType(healthcareEntityRelation1,
+            "QualifierOfCondition");
+        HealthcareEntityRelationPropertiesHelper.setBidirectional(healthcareEntityRelation1, false);
+        HealthcareEntityRelationPropertiesHelper.setSourceLink(healthcareEntityRelation1,
+            "#/results/documents/0/entities/2");
+        HealthcareEntityRelationPropertiesHelper.setTargetLink(healthcareEntityRelation1,
+            "#/results/documents/0/entities/3");
+
+        final HealthcareEntityRelation healthcareEntityRelation2 = new HealthcareEntityRelation();
+        HealthcareEntityRelationPropertiesHelper.setRelationType(healthcareEntityRelation2, "TimeOfCondition");
+        HealthcareEntityRelationPropertiesHelper.setBidirectional(healthcareEntityRelation2, false);
+        HealthcareEntityRelationPropertiesHelper.setSourceLink(healthcareEntityRelation2,
+            "#/results/documents/0/entities/4");
+        HealthcareEntityRelationPropertiesHelper.setTargetLink(healthcareEntityRelation2,
+            "#/results/documents/0/entities/3");
+
+        final HealthcareEntityRelation healthcareEntityRelation3 = new HealthcareEntityRelation();
+        HealthcareEntityRelationPropertiesHelper.setRelationType(healthcareEntityRelation3, "TimeOfCondition");
+        HealthcareEntityRelationPropertiesHelper.setBidirectional(healthcareEntityRelation3, false);
+        HealthcareEntityRelationPropertiesHelper.setSourceLink(healthcareEntityRelation3,
+            "#/results/documents/1/entities/0");
+        HealthcareEntityRelationPropertiesHelper.setTargetLink(healthcareEntityRelation3,
+            "#/results/documents/1/entities/2");
+
+        final HealthcareEntityRelation healthcareEntityRelation4 = new HealthcareEntityRelation();
+        HealthcareEntityRelationPropertiesHelper.setRelationType(healthcareEntityRelation4,
+            "QualifierOfCondition");
+        HealthcareEntityRelationPropertiesHelper.setBidirectional(healthcareEntityRelation4, false);
+        HealthcareEntityRelationPropertiesHelper.setSourceLink(healthcareEntityRelation4,
+            "#/results/documents/1/entities/1");
+        HealthcareEntityRelationPropertiesHelper.setTargetLink(healthcareEntityRelation4,
+            "#/results/documents/1/entities/2");
+
+        final HealthcareEntityCollection healthcareEntityCollection1 = new HealthcareEntityCollection(
+            new IterableStream<>(asList(healthcareEntity1, healthcareEntity2, healthcareEntity3, healthcareEntity4,
+                healthcareEntity5)));
+        HealthcareEntityCollectionPropertiesHelper.setEntityRelations(healthcareEntityCollection1,
+            new IterableStream<>(asList(healthcareEntityRelation1, healthcareEntityRelation2)));
+
+        final HealthcareEntityCollection healthcareEntityCollection2 = new HealthcareEntityCollection(
+            new IterableStream<>(asList(healthcareEntity6, healthcareEntity7, healthcareEntity8, healthcareEntity9,
+                healthcareEntity10, healthcareEntity11)));
+        HealthcareEntityCollectionPropertiesHelper.setEntityRelations(healthcareEntityCollection2,
+            new IterableStream<>(asList(healthcareEntityRelation3, healthcareEntityRelation4)));
+
+        final RecognizeHealthcareEntitiesResult healthcareEntitiesResult1 = new RecognizeHealthcareEntitiesResult("0",
+            textDocumentStatistics1, null);
+        RecognizeHealthcareEntitiesResultPropertiesHelper.setEntities(healthcareEntitiesResult1,
+            healthcareEntityCollection1);
+
+        final RecognizeHealthcareEntitiesResult healthcareEntitiesResult2 = new RecognizeHealthcareEntitiesResult("1",
+            textDocumentStatistics2, null);
+        RecognizeHealthcareEntitiesResultPropertiesHelper.setEntities(healthcareEntitiesResult2,
+            healthcareEntityCollection2);
+
+        final RecognizeHealthcareEntitiesResultCollection recognizeHealthcareEntitiesResults =
+            new RecognizeHealthcareEntitiesResultCollection(asList(healthcareEntitiesResult1,
+                healthcareEntitiesResult2));
+        RecognizeHealthcareEntitiesResultCollectionPropertiesHelper.setModelVersion(recognizeHealthcareEntitiesResults,
+            "2020-09-03");
+        RecognizeHealthcareEntitiesResultCollectionPropertiesHelper.setStatistics(recognizeHealthcareEntitiesResults,
+            textDocumentBatchStatistics);
+
+        return recognizeHealthcareEntitiesResults;
     }
 
     /**